🏝️  feat: update ts error for login component
