+++ src/modules/Project/data/index.ts
@@ -22,7 +22,7 @@ export const projectList = [
     corpName: '测试公司测试公司测试公司测试公司测试公司测试公司测试公司测试公司测试公司测试公司测试公司测试公司测试公司',
     notes: '备注备注',
     isPublished: true,
-    createTime: '2021.03.29'
+    createTime: '2023-10-01'
   },
   {
     id: '333',
@@ -30,7 +30,7 @@ export const projectList = [
     corpName: '测试公司',
     notes: '',
     isPublished: false,
-    createTime: '2021.02.29'
+    createTime: '2023-10-01'
   }
 ]
