+++ src/types/global.d.ts
@@ -4,3 +4,30 @@ interface Window {
   $ModalDialog: import('naive-ui').DialogProviderInst
   $ModalLoadingBar: import('naive-ui').LoadingBarProviderInst
 }
+
+declare namespace FormContainer {
+  type Attrs = {
+    path: string
+    rule: () => any
+  }
+
+  interface FormConfigOptions {
+    attrs: Attrs
+    link?: any
+    type?: string
+    label: string
+    prefixIcon?: Component
+    showPasswordOn?: string
+    placeholder: string
+  }
+
+  type ActionOnEvent = {
+    click: (...args: any[]) => any
+  }
+  interface ActionOptions {
+    attrs: any
+    text: string
+    on: ActionOnEvent
+  }
+
+}
